+++ src/promptflow/CHANGELOG.md
@@ -8,6 +8,7 @@
 - [SDK/CLI] Support custom strong type connection.
 - [SDK/CLI] Enable telemetry and won't collect by default, use `pf config set cli.telemetry_enabled=true` to opt in.
 - [SDK/CLI] Exposed function `from promptflow import load_run` to load run object from local YAML file.
+- [Executor] Support `ToolProvider` for script tools.
 
 ### Bugs Fixed
 - **pf config set**: